Replaced config.yml example Added meta.yml that gives algorithm options Removed more traces of redis Made work of each task run in temp directory changed constants to enums Used typing module in places that it should have been used in the first place significant changes to space_settings to eliminate getattr and instead return access normally Settings objects renamed based on Task Project task separated from calculate sim since not all jobs require sims be calculated Stub for convert task created, doesn't work yet echo_message moved to newly created utils file case_insensitive changed to case_sensitive based on negation confusing people File created for project task sim_batch_worker removed sim_calculator replaced with sim_calculator3 internal storage mechanism changed to h5 vectorizer renamed to vec_worker since it's only used in pools last traces of redis removed from entrypoint migrated from alpine to debian

+++ py/vec_worker.py
@@ -10,7 +10,7 @@ class Vectorizer(object):
     def vectorize(self, key, document):
         bow = self.dictionary.doc2bow(document)
         ms = self.model[bow]
-        return key, np.array([n[1] for n in ms])
+        return key, np.array([n[1] for n in ms], dtype=np.float32)
 
 v: Vectorizer
